+++ executor/counters.go
