+++ executor/executor.go
@@ -2,7 +2,9 @@ package executor
 
 import (
 	"bytes"
+	"context"
 	"encoding/json"
+	"errors"
 	"flag"
 	"os"
 	"sort"
@@ -12,14 +14,15 @@ import (
 
 	"github.com/naturali/kmr/bucket"
 	"github.com/naturali/kmr/job"
+	"github.com/naturali/kmr/mapred"
 	"github.com/naturali/kmr/master"
 	kmrpb "github.com/naturali/kmr/pb"
 	"github.com/naturali/kmr/records"
 	"github.com/naturali/kmr/util"
 	"github.com/naturali/kmr/util/log"
 
-	"golang.org/x/net/context"
 	"google.golang.org/grpc"
+	"github.com/reusee/mmh3"
 )
 
 var (
@@ -37,193 +40,38 @@ var (
 	masterAddr = flag.String("master-addr", "", "the address of master")
 )
 
-type ComputeWrap struct {
-	mapFunc     func(kvs <-chan *kmrpb.KV) <-chan *kmrpb.KV
-	reduceFunc  func(kvs <-chan *kmrpb.KV) <-chan *kmrpb.KV
+const (
+	ReducerConcurrentLevel = 4
+)
+
+func getCollectFunc(outClass mapred.OutputClassBase, outputKV chan<- *kmrpb.KV) func(k, v interface{}) {
+	return func(k, v interface{}) {
+		keyBytes := outClass.GetOutputKeyClass().ToBytes(k)
+		valueBytes := outClass.GetOutputValueClass().ToBytes(v)
+		outputKV <- &kmrpb.KV{Key: keyBytes, Value: valueBytes}
+	}
+}
+
+type ComputeWrapClass struct {
+	mapClass    mapred.Mapper
+	reduceClass mapred.Reducer
+	// combineClass mapred.Reducer
 	combineFunc func(v1 []byte, v2 []byte) []byte
 }
 
-func (cw *ComputeWrap) BindMapper(mapper func(kvs <-chan *kmrpb.KV) <-chan *kmrpb.KV) {
-	cw.mapFunc = mapper
+func (cw *ComputeWrapClass) BindMapper(mapper mapred.Mapper) {
+	cw.mapClass = mapper
 }
 
-func (cw *ComputeWrap) BindReducer(reducer func(kvs <-chan *kmrpb.KV) <-chan *kmrpb.KV) {
-	cw.reduceFunc = reducer
+func (cw *ComputeWrapClass) BindReducer(reducer mapred.Reducer) {
+	cw.reduceClass = reducer
 }
 
-func (cw *ComputeWrap) BindCombiner(combiner func(v1 []byte, v2 []byte) []byte) {
+func (cw *ComputeWrapClass) BindCombiner(combiner func(v1 []byte, v2 []byte) []byte) {
 	cw.combineFunc = combiner
 }
 
-func (cw *ComputeWrap) Run() {
-	flag.Parse()
-
-	if os.Getenv("KMR_MASTER_ADDRESS") != "" {
-		addr := os.Getenv("KMR_MASTER_ADDRESS")
-		masterAddr = &addr
-	}
-
-	if *masterAddr == "" {
-		// Local Run
-		var taskID int
-		switch *phase {
-		case "map":
-			taskID = *mapID
-		case "reduce":
-			taskID = *reduceID
-		}
-		bucket := job.BucketDescription{
-			BucketType: "filesystem",
-			Config:     job.BucketConfig{"directory": *dataDir},
-		}
-		err := cw.phaseSelector(*jobName, *phase,
-			bucket, bucket, bucket,
-			[]string{*inputFile}, *nMap, *nReduce, *readerType, taskID, 0, make([]int64, *nMap), 1)
-		if err != nil {
-			log.Fatal(err)
-		}
-	} else {
-		var retcode kmrpb.ReportInfo_ErrorCode
-		// Distributed Mode
-		cc, err := grpc.Dial(*masterAddr, grpc.WithInsecure())
-		if err != nil {
-			log.Fatal("cannot connect to master", err)
-		}
-		masterClient := kmrpb.NewMasterClient(cc)
-		for {
-			task, err := masterClient.RequestTask(context.Background(), &kmrpb.RegisterParams{
-				JobName: *jobName,
-			})
-			if err != nil || task.Retcode != 0 {
-				log.Error(err)
-				// TODO: random backoff
-				time.Sleep(1 * time.Second)
-				continue
-			}
-			taskInfo := task.Taskinfo
-			timer := time.NewTicker(master.HEARTBEAT_TIMEOUT / 2)
-			go func() {
-				for range timer.C {
-					// SendHeartBeat
-					masterClient.ReportTask(context.Background(), &kmrpb.ReportInfo{
-						JobName:  *jobName,
-						Phase:    taskInfo.Phase,
-						TaskID:   taskInfo.TaskID,
-						WorkerID: task.WorkerID,
-						Retcode:  kmrpb.ReportInfo_DOING,
-					})
-				}
-			}()
-			var mapBucket, interBucket, reduceBucket job.BucketDescription
-			json.Unmarshal([]byte(taskInfo.MapBucketJson), &mapBucket)
-			json.Unmarshal([]byte(taskInfo.IntermediateBucketJson), &interBucket)
-			json.Unmarshal([]byte(taskInfo.ReduceBucketJson), &reduceBucket)
-
-			err = cw.phaseSelector(taskInfo.JobName, taskInfo.Phase,
-				mapBucket, interBucket, reduceBucket, taskInfo.Files,
-				int(taskInfo.NMap), int(taskInfo.NReduce), taskInfo.ReaderType, int(taskInfo.TaskID), task.WorkerID, taskInfo.CommitMappers, int(taskInfo.MapBatchSize))
-			retcode = kmrpb.ReportInfo_FINISH
-			if err != nil {
-				log.Debug(err)
-				retcode = kmrpb.ReportInfo_ERROR
-			}
-			timer.Stop()
-			masterClient.ReportTask(context.Background(), &kmrpb.ReportInfo{
-				JobName:  *jobName,
-				Phase:    taskInfo.Phase,
-				TaskID:   taskInfo.TaskID,
-				WorkerID: task.WorkerID,
-				Retcode:  retcode,
-			})
-			// backoff
-			if err != nil {
-				time.Sleep(1 * time.Second)
-			}
-		}
-	}
-}
-
-func (cw *ComputeWrap) phaseSelector(jobName string, phase string,
-	mapBucket, interBucket, reduceBucket job.BucketDescription, files []string,
-	nMap int, nReduce int, readerType string, taskID int, workerID int64,
-	commitMappers []int64, mapBatchSize int) error {
-	interBk, err := bucket.NewBucket(interBucket.BucketType, interBucket.Config)
-	if err != nil {
-		log.Fatalf("Fail to open bucket: %v", err)
-	}
-	switch phase {
-	case "map":
-		log.Infof("starting id%d mapper", taskID)
-		mapBk, err := bucket.NewBucket(mapBucket.BucketType, mapBucket.Config)
-		if err != nil {
-			log.Fatalf("Fail to open mapbucket: %v", err)
-		}
-		readers := make([]records.RecordReader, 0)
-		for _, file := range files {
-			reader, err := mapBk.OpenRead(file)
-			if err != nil {
-				log.Fatalf("Fail to open object %s: %v", file, err)
-			}
-			recordReader := records.MakeRecordReader(readerType, map[string]interface{}{"reader": reader})
-			readers = append(readers, recordReader)
-		}
-		batchReader := records.NewChainReader(readers)
-		// Mapper
-		if err := cw.doMap(batchReader, interBk, taskID, nReduce, workerID); err != nil {
-			log.Fatalf("Fail to Map: %v", err)
-		}
-		batchReader.Close()
-	case "reduce":
-		log.Infof("starting id%d reducer", nReduce)
-		reduceBk, err := bucket.NewBucket(reduceBucket.BucketType, reduceBucket.Config)
-		if err != nil {
-			log.Fatalf("Fail to open reducebucket: %v", err)
-		}
-		// Reduce
-		outputFile := "res-" + strconv.Itoa(taskID) + ".t"
-		writer, err := reduceBk.OpenWrite(outputFile)
-		if err != nil {
-			log.Fatalf("Failed to open intermediate: %v", err)
-		}
-		recordWriter := records.MakeRecordWriter("stream", map[string]interface{}{"writer": writer})
-		if err = cw.doReduce(interBk, taskID, nMap, mapBatchSize, commitMappers, recordWriter); err != nil {
-			log.Fatalf("Fail to Reduce: %v", err)
-		}
-		recordWriter.Close()
-	default:
-		panic("bad phase")
-	}
-	log.Info("Exit executor")
-	return nil
-}
-
-func (cw *ComputeWrap) sortAndCombine(aggregated []*records.Record) []*records.Record {
-	sort.Slice(aggregated, func(i, j int) bool {
-		return bytes.Compare(aggregated[i].Key, aggregated[j].Key) < 0
-	})
-	if cw.combineFunc == nil {
-		return aggregated
-	}
-	combined := make([]*records.Record, 0)
-	curRecord := &records.Record{}
-	for _, r := range aggregated {
-		if !bytes.Equal(curRecord.Key, r.Key) {
-			if curRecord.Key != nil {
-				combined = append(combined, curRecord)
-			}
-			curRecord = r
-		} else {
-			curRecord.Value = cw.combineFunc(curRecord.Value, r.Value)
-		}
-	}
-	if curRecord.Key != nil {
-		combined = append(combined, curRecord)
-	}
-	return combined
-}
-
-// doMap does map operation and save the intermediate files.
-func (cw *ComputeWrap) doMap(rr records.RecordReader, bk bucket.Bucket, mapID int, nReduce int, workerID int64) (err error) {
+func (cw *ComputeWrapClass) doMap(rr records.RecordReader, bk bucket.Bucket, mapID int, nReduce int, workerID int64) (err error) {
 	maxNumConcurrentFlush := 2
 	startTime := time.Now()
 	aggregated := make([]*records.Record, 0)
@@ -233,14 +81,27 @@ func (cw *ComputeWrap) doMap(rr records.RecordReader, bk bucket.Bucket, mapID in
 	// map
 	waitc := make(chan struct{})
 	inputKV := make(chan *kmrpb.KV, 1024)
-	outputKV := cw.mapFunc(inputKV)
+	outputKV := make(chan *kmrpb.KV, 1024)
+	// outputKV := cw.mapFunc(inputKV)
+	cw.mapClass.BeforeRun()
+	keyClass, valueClass := cw.mapClass.GetInputKeyClass(), cw.mapClass.GetInputValueClass()
+	go func() {
+		for kvpair := range inputKV {
+			key := keyClass.FromBytes(kvpair.Key)
+			value := valueClass.FromBytes(kvpair.Value)
+			//TODO: implement report
+			cw.mapClass.Map(key, value, getCollectFunc(cw.mapClass, outputKV), nil)
+		}
+		close(outputKV)
+	}()
 	go func() {
 		sem := util.NewSemaphore(maxNumConcurrentFlush)
 		var waitFlushWrite sync.WaitGroup
 		for in := range outputKV {
 			aggregated = append(aggregated, KVToRecord(in))
 			currentAggregatedSize += 8 + len(in.Key) + len(in.Value)
-			if currentAggregatedSize >= *flushSize*1024*1024 {
+			// if currentAggregatedSize >= *flushSize*1024*1024 {
+			if currentAggregatedSize >= 1*1024*1024 {
 				filename := bucket.FlushoutFileName("map", mapID, len(flushOutFiles), workerID)
 				sem.Acquire(1)
 				go func(filename string, data []*records.Record) {
@@ -333,7 +194,7 @@ func (cw *ComputeWrap) doMap(rr records.RecordReader, bk bucket.Bucket, mapID in
 }
 
 // doReduce does reduce operation
-func (cw *ComputeWrap) doReduce(interBk bucket.Bucket, reduceID int, nMap, batchSize int, commitMappers []int64, writer records.RecordWriter) error {
+func (cw *ComputeWrapClass) doReduce(interBk bucket.Bucket, reduceID int, nMap, batchSize int, commitMappers []int64, writer records.RecordWriter) error {
 	startTime := time.Now()
 	readers := make([]records.RecordReader, 0)
 	nBatch := nMap / batchSize
@@ -351,9 +212,48 @@ func (cw *ComputeWrap) doReduce(interBk bucket.Bucket, reduceID int, nMap, batch
 	sorted := make(chan *records.Record, 1024)
 	go records.MergeSort(readers, sorted)
 
-	inputs := make(chan *kmrpb.KV, 1024)
-	outputs := cw.reduceFunc(inputs)
+	outputs := make(chan *kmrpb.KV, 1024)
+	cw.reduceClass.BeforeRun()
+	reducerCaller := func(inputs <-chan *kmrpb.KV, wg *sync.WaitGroup) {
+		keyClass, valueClass := cw.reduceClass.GetInputKeyClass(), cw.reduceClass.GetInputValueClass()
+		startKVPair := <-inputs
+		stopFlag := startKVPair == nil
+		for !stopFlag {
+			key := keyClass.FromBytes(startKVPair.Key)
+			curKeyBytes := startKVPair.Key[:]
+			reduceIteratedAllValue := false
+			nextFunc := func() (interface{}, error) {
+				if reduceIteratedAllValue {
+					return nil, errors.New("EOINPUT")
+				}
+				if startKVPair != nil {
+					tmp := startKVPair.Value[:]
+					startKVPair = nil
+					return valueClass.FromBytes(tmp), nil
+				}
+				kvpair := <-inputs
+				if kvpair != nil {
+					if bytes.Equal(kvpair.Key[:], curKeyBytes[:]) {
+						return valueClass.FromBytes(kvpair.Value), nil
+					}
+					startKVPair = kvpair
+					reduceIteratedAllValue = true
+					return nil, errors.New("EOKEY")
+				}
+				stopFlag = true
+				reduceIteratedAllValue = true
+				return nil, errors.New("EOINPUT")
+			}
+			cw.reduceClass.Reduce(key, nextFunc, getCollectFunc(cw.reduceClass, outputs), nil)
+			for !reduceIteratedAllValue {
+				nextFunc()
+			}
+		}
+		wg.Done()
+	}
+
 	var wg sync.WaitGroup
+	var wgForReducerCaller sync.WaitGroup
 	wg.Add(1)
 	go func(outputs <-chan *kmrpb.KV) {
 		for r := range outputs {
@@ -361,12 +261,204 @@ func (cw *ComputeWrap) doReduce(interBk bucket.Bucket, reduceID int, nMap, batch
 		}
 		wg.Done()
 	}(outputs)
-	for r := range sorted {
-		inputs <- RecordToKV(r)
+
+	var inputsArr [ReducerConcurrentLevel]chan *kmrpb.KV
+	for i := range inputsArr {
+		inputsArr[i] = make(chan *kmrpb.KV, 1024)
+		wgForReducerCaller.Add(1)
+		go reducerCaller(inputsArr[i], &wgForReducerCaller)
 	}
-	close(inputs)
+
+	for r := range sorted {
+		if ReducerConcurrentLevel == 1 {
+			inputsArr[0] <- RecordToKV(r)
+		} else {
+			k := mmh3.Hash32(r.Key[:]) % ReducerConcurrentLevel
+			inputsArr[k] <- RecordToKV(r)
+		}
+	}
+
+	for i := range inputsArr {
+		inputsArr[i] <- nil
+		close(inputsArr[i])
+	}
+
+	wgForReducerCaller.Wait()
+	close(outputs)
+
 	wg.Wait()
 	log.Debug("DONE Reduce. Took:", time.Since(startTime))
+	return nil
+}
+
+func (cw *ComputeWrapClass) sortAndCombine(aggregated []*records.Record) []*records.Record {
+	sort.Slice(aggregated, func(i, j int) bool {
+		return bytes.Compare(aggregated[i].Key, aggregated[j].Key) < 0
+	})
+	if cw.combineFunc == nil {
+		return aggregated
+	}
+	combined := make([]*records.Record, 0)
+	curRecord := &records.Record{}
+	for _, r := range aggregated {
+		if !bytes.Equal(curRecord.Key, r.Key) {
+			if curRecord.Key != nil {
+				combined = append(combined, curRecord)
+			}
+			curRecord = r
+		} else {
+			curRecord.Value = cw.combineFunc(curRecord.Value, r.Value)
+		}
+	}
+	if curRecord.Key != nil {
+		combined = append(combined, curRecord)
+	}
+	return combined
+}
+
+func (cw *ComputeWrapClass) aggregateValues(inputs <-chan *kmrpb.KV, values chan<- interface{}) {
+
+}
+
+func (cw *ComputeWrapClass) Run() {
+	flag.Parse()
+
+	if os.Getenv("KMR_MASTER_ADDRESS") != "" {
+		addr := os.Getenv("KMR_MASTER_ADDRESS")
+		masterAddr = &addr
+	}
+
+	if *masterAddr == "" {
+		// Local Run
+		var taskID int
+		switch *phase {
+		case "map":
+			taskID = *mapID
+		case "reduce":
+			taskID = *reduceID
+		}
+		bucket := job.BucketDescription{
+			BucketType: "filesystem",
+			Config:     job.BucketConfig{"directory": *dataDir},
+		}
+		err := cw.phaseSelector(*jobName, *phase,
+			bucket, bucket, bucket,
+			[]string{*inputFile}, *nMap, *nReduce, *readerType, taskID, 0, make([]int64, *nMap), 1)
+		if err != nil {
+			log.Fatal(err)
+		}
+	} else {
+		var retcode kmrpb.ReportInfo_ErrorCode
+		// Distributed Mode
+		cc, err := grpc.Dial(*masterAddr, grpc.WithInsecure())
+		if err != nil {
+			log.Fatal("cannot connect to master", err)
+		}
+		masterClient := kmrpb.NewMasterClient(cc)
+		for {
+			task, err := masterClient.RequestTask(context.Background(), &kmrpb.RegisterParams{
+				JobName: *jobName,
+			})
+			if err != nil || task.Retcode != 0 {
+				log.Error(err)
+				// TODO: random backoff
+				time.Sleep(1 * time.Second)
+				continue
+			}
+			taskInfo := task.Taskinfo
+			timer := time.NewTicker(master.HEARTBEAT_TIMEOUT / 2)
+			go func() {
+				for range timer.C {
+					// SendHeartBeat
+					masterClient.ReportTask(context.Background(), &kmrpb.ReportInfo{
+						JobName:  *jobName,
+						Phase:    taskInfo.Phase,
+						TaskID:   taskInfo.TaskID,
+						WorkerID: task.WorkerID,
+						Retcode:  kmrpb.ReportInfo_DOING,
+					})
+				}
+			}()
+			var mapBucket, interBucket, reduceBucket job.BucketDescription
+			json.Unmarshal([]byte(taskInfo.MapBucketJson), &mapBucket)
+			json.Unmarshal([]byte(taskInfo.IntermediateBucketJson), &interBucket)
+			json.Unmarshal([]byte(taskInfo.ReduceBucketJson), &reduceBucket)
+
+			err = cw.phaseSelector(taskInfo.JobName, taskInfo.Phase,
+				mapBucket, interBucket, reduceBucket, taskInfo.Files,
+				int(taskInfo.NMap), int(taskInfo.NReduce), taskInfo.ReaderType, int(taskInfo.TaskID), task.WorkerID, taskInfo.CommitMappers, int(taskInfo.MapBatchSize))
+			retcode = kmrpb.ReportInfo_FINISH
+			if err != nil {
+				log.Debug(err)
+				retcode = kmrpb.ReportInfo_ERROR
+			}
+			timer.Stop()
+			masterClient.ReportTask(context.Background(), &kmrpb.ReportInfo{
+				JobName:  *jobName,
+				Phase:    taskInfo.Phase,
+				TaskID:   taskInfo.TaskID,
+				WorkerID: task.WorkerID,
+				Retcode:  retcode,
+			})
+			// backoff
+			if err != nil {
+				time.Sleep(1 * time.Second)
+			}
+		}
+	}
+}
+
+func (cw *ComputeWrapClass) phaseSelector(jobName string, phase string,
+	mapBucket, interBucket, reduceBucket job.BucketDescription, files []string,
+	nMap int, nReduce int, readerType string, taskID int, workerID int64,
+	commitMappers []int64, mapBatchSize int) error {
+	interBk, err := bucket.NewBucket(interBucket.BucketType, interBucket.Config)
+	if err != nil {
+		log.Fatalf("Fail to open bucket: %v", err)
+	}
+	switch phase {
+	case "map":
+		log.Infof("starting id%d mapper", taskID)
+		mapBk, err := bucket.NewBucket(mapBucket.BucketType, mapBucket.Config)
+		if err != nil {
+			log.Fatalf("Fail to open mapbucket: %v", err)
+		}
+		readers := make([]records.RecordReader, 0)
+		for _, file := range files {
+			reader, err := mapBk.OpenRead(file)
+			if err != nil {
+				log.Fatalf("Fail to open object %s: %v", file, err)
+			}
+			recordReader := records.MakeRecordReader(readerType, map[string]interface{}{"reader": reader})
+			readers = append(readers, recordReader)
+		}
+		batchReader := records.NewChainReader(readers)
+		// Mapper
+		if err := cw.doMap(batchReader, interBk, taskID, nReduce, workerID); err != nil {
+			log.Fatalf("Fail to Map: %v", err)
+		}
+		batchReader.Close()
+	case "reduce":
+		log.Infof("starting id%d reducer", nReduce)
+		reduceBk, err := bucket.NewBucket(reduceBucket.BucketType, reduceBucket.Config)
+		if err != nil {
+			log.Fatalf("Fail to open reducebucket: %v", err)
+		}
+		// Reduce
+		outputFile := "res-" + strconv.Itoa(taskID) + ".t"
+		writer, err := reduceBk.OpenWrite(outputFile)
+		if err != nil {
+			log.Fatalf("Failed to open intermediate: %v", err)
+		}
+		recordWriter := records.MakeRecordWriter("stream", map[string]interface{}{"writer": writer})
+		if err = cw.doReduce(interBk, taskID, nMap, mapBatchSize, commitMappers, recordWriter); err != nil {
+			log.Fatalf("Fail to Reduce: %v", err)
+		}
+		recordWriter.Close()
+	default:
+		panic("bad phase")
+	}
+	log.Info("Exit executor")
 	return nil
 }
